Added leadership to implicit selectors. (#11)
Added leadership to implicit selectors.

To make this work, I had to refactor the way that we do the fetching and
composing of units for the implicit selectors. It's not as elegant or
flexible any more, but it does support asynchrounous operations, which
was critical.

I also had to make selectors asynchronous, and get rid of type checking
for their returns -- verifying that they all return a coroutine object
isn't as interesting as being able to verify what that coroutine
actually returns ...

Also fixed incorrect indentation in code involving waiters. Got rid of some
extraneous logging messages. Incorporated a fix from python-libjuju.

And added catch for exception when running pkill jujud.

Putting the catch here rather than in libjuju because we are doing
something exceptional (killing the juju daemon), and it's probably
appropriate for python-libjuju to complain.