Only run against local bundles, and merge tests by name
We'll rely on a higher level (bundletester) to fetch bundles from the
store, so we can assume we're always running against a local bundle.
We also want to merge tests by name, and if the name is already present,
it will be overwritten by the new test definition.  Also, support
pulling in additional suites besides the default suite and the bundle
suite, with the option to suppress the default or bundle suites entirely.