Support tests and tasks from bundle under test (#26)
* Made bundle required CLI param instead of part of test plan

Also added default test plan

* Replace model and debug-log view with Terminal to support colors

* Support tests and tasks from bundle under test

Discover a matrix.yaml to be merged with the selected test plan from the
bundle being tested.  Also include that bundle in the path so that
custom tasks can be used.

The libjuju changes are for deploying local bundles and fixing placement.

* Only run against local bundles, and merge tests by name

We'll rely on a higher level (bundletester) to fetch bundles from the
store, so we can assume we're always running against a local bundle.
We also want to merge tests by name, and if the name is already present,
it will be overwritten by the new test definition.  Also, support
pulling in additional suites besides the default suite and the bundle
suite, with the option to suppress the default or bundle suites entirely.

* Cleanup superfluous PyYAML wheel

* Update libjuju to fix websocket error and minor change to default suite

* Added usage examples to --help

* Added additional usage example