Skip subordinate by default on actions.
Most of our actions don't really apply to subordinate charms, so we skip
subordinate charms by default when putting a test plan together.

There are some actions, like "upgrade" that do make sense for a
subordinate charm, so I added a method of "tagging" actions with
subordinate okay. The tagging system can also be expanded to handle
other similar circumstances in the future.