Reactivated juju-crashdump.
We skip it if it doesn't exist, but run it if it does.

We also now skip packing the matrix.log and glitch_plan into a tarball,
as that can fail if we are logging while doing it, and it can be handled
on a higher level.

Tweaked tests so that they don't expect matrix tarballs (or crashdump
tarballs.)