Tweaks for versioning and packaging. (#98)
Tweaks for versioning and packaging.

`python setup.py sdist` now produces a tarball that works with pip
install. (More ways to install!)

Updated version (we haven't done so in a while).

Got rid of "this is an alpha!" scare text. We're now in beta :-)

Changed executable to `juju-matrix`, to be consistent with plans for the
snap. This has the added benefit of making matrix work as a juju
"plugin" -- i.e., it can be run as `juju matrix`.

(Also temporarily have the old script to ease the transition.)

Changed name of package in setup.py to jujumatrix, to reflect our plans
for posting to pypi without conflicting with the existing "matrix" package.