Fixed info spam on model deploy failure.
Added a default=None for juju_model, which allows the check for
"juju_model=None" to succeed in the teardown. This prevents the case
where you get a lot of "Attribute has no attribute info" spam in the
logs when attempting to cleanup a test where the model failed to deploy.