Added conjure-up support. (#107)
This is basic support for calling out to conjure-up on the command line,
and having it handle the deploy. It requires that you have a version of
conjure-up in your environment that has the patch referenced in
https://github.com/conjure-up/conjure-up/pull/760

Matrix will use automagic to determine what method it should use to deploy.
It will only use conjure-up when pointed at a spell.

Also note that Travis can't install the conjure-up snap, so functional tests
for conjure-up are skipped in Travis.

Finally, I took a moment to improve the functional testing logging. 
Functional tests no longer sit there silently staring at you while they do stuff!