Removed SUBORDINATE_OK flag from reboot and kill_juju_agent
The flag wasn't working consistently, and it's not really necessary to
automatically generate reboot and kill_juju_agent actions for
subordinate charms. There's always a parent charm that we can run those
actions against, instead.

Leaving the SUBORDINATE_OK flag in the codebase for now, just in case we
do add an action that needs it.