Matrix now fails rather than hangs when disconnected.
If the websocket connection to the Juju controller was broken while a
matrix test was running, the health check could become stuck in a "busy"
state, never receiving updates that told it to proceed. This caused
matrix to run for a long time, finally failing by timing out.

Now, matrix detects that the websocket connection has gone away, and
raises an InfraFailure, terminating the test with an exit code of 1 as
soon as it detects the connection failure.