Updated python-libjuju, and fixed connection error detection. (#131)
We no longer need to attach to the Monitor class, as exceptions now get
propagated. Added a hook to set exit_code to 1 in the cases where we
have an unhandled Exception.

Updated with libjuju 0.4.1 release binary.

Got rid of spelunking in logs for connection failure in connection_test -- not reliable enough.